se agrego cambio de roles y owner

+++ src/routes/products.router.js
@@ -33,17 +33,30 @@ router.get("/:id", async (req, res) => {
 
 //DELETE
 router.delete("/:pid", async (req, res) => {
-    const id = req.params.pid
-    const productDeleted = await ProductService.delete(id)
-
-    req.io.emit('updatedProducts', await ProductService.get());
-    res.json({status: "Success", massage: "Product Deleted!", productDeleted})
+        const id = req.params.pid
+        const user = req.user.user
+        const productID = await ProductService.getOneByID(id)
+        const owner = productID.owner.id
+        if(owner == user._id || user.role == "admin"){
+            console.log("si")
+            const productDeleted = await ProductService.delete(id) 
+            req.io.emit('updatedProducts', await ProductService.get());
+            res.json({status: "Success", massage: "Product Deleted!", productDeleted})
+    }else{
+        req.logger.warning("No Owner")
+    }
 })
 
 //POST
 router.post("/", async (req, res) => {
     try {
+        const user = req.user.user
         const product = req.body
+        product.owner = {
+            role: user.role,
+            email: user.email,
+            id: user._id
+        }
         if (!product.title) {
             return res.status(400).json({message: "Error Falta el nombre del producto"})
         }
